Start working on the new egui-free graph data model
The plan is to make the core of blackjack more detached from
blackjack_ui. The role of blackjack_ui will be to provide a user
interface, in a way that the state of the user interface can be
converted to this data model (and then, compiled to Lua) at any time.