Start working on stable serialization format.
This is a small detour initiated in #53, due to the fact that the Jack
file format is now broken after we no longer represent graphs themselves
as compiled lua code (which is a prerequisite to implement Gizmos, among
other useful optimizations).

The idea is to get a single canonical file format that can be used both
as an editor representation, and a representation for exported 'Jacks'.
Additionally, we will take extra care to make the format
forwards-compatible by defining mirror structs for the serialization.
This may be tedious, but gives us better control of the serialization
output and lets us build a more future-proof format.