Remove BlackjackParameter from the codebase
Also remove WIP serialization code for it. The rationale is that we
don't want to store parameter configuration in the graph or the nodes.
This is data that corresponds to the nodes, and is stored in the Lua
code describing each node. Having it duplicated was always a data
redundancy waiting to bite us!