Add WIP engine tests.
The idea is to load bjk files, run the graph, and then assert some
properties of the final mesh.

The idea is that these tests will break if the stored files from the
examples ever stop running, so it will be the best reminder to keep
examples up to date.

This should also ensure that the bjk data format is backwards
compatible. Or at least will make it very clear when there are breaking
changes.